Added View to Show how Modals work with SwiftUI
The added button of the image displays a modal with a detail view of the selected image. A custom dismiss button is added with an alert presenting to ensure the user is done with the page they are viewing. This is done by removing Navigation View and creating a custom action to dimiss the model with in the alert